refactor: rewrite all_warnings_view to be stateless and easier to follow
All logic of retrieving warnings is now done in providers and not in the
UI any more, this allows the UI to be stateless.
I've also removed leftovers of an old source of alerts which we don't
use any more since we switched to FPAS.

We probably want to poll for warnings in the future so the user doesn't
need to manually refresh/reload, but this is a good start to continue
working on.